pyproject.toml
[project]
name = "rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "distrax>=0.1.5",
    "flax>=0.10.4",
    "hydra-core>=1.3.2",
    "jax==0.5.0",
    "jax-metal==0.1.1",
    "jaxlib==0.5.0",
    "optax>=0.2.5",
    "pgx>=2.6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]
